+++ fsvolumelist/listwindow.cpp
@@ -209,9 +209,16 @@ public:
 	virtual VOID InitLayout(const RECT *prcDesktopWorkArea)
 	{
 		RECT rc;
-		GetClientRect(m_hWnd,&rc);
-		UpdateLayout(_RECT_WIDTH(rc),_RECT_HIGHT(rc),FALSE);
-		m_pView->InitLayout(&rc);
+		if( prcDesktopWorkArea )
+		{
+			rc = *prcDesktopWorkArea;
+		}
+		else
+		{
+			GetClientRect(m_hWnd,&rc);
+			UpdateLayout(_RECT_WIDTH(rc),_RECT_HIGHT(rc),FALSE);
+			m_pView->InitLayout(&rc);
+		}
 		m_hWndCtrlFocus = m_pView->GetPageHWND();
 	}
 };